Refactoring the repositories' interfaces location, now those are in the application project.
