+++ src/3ASystem.Application/Abstractions/Data/Repositories/IFunctionalityRepository.cs
@@ -1,0 +1,7 @@
+﻿using _3ASystem.Domain.Entities.Functionalities;
+
+namespace _3ASystem.Application.Abstractions.Data.Repositories;
+
+public interface IFunctionalityRepository : IRepository<Functionality, FunctionalityId>
+{
+}
