+++ src/3ASystem.Application/Abstractions/Data/Repositories/IOperationRepository.cs
@@ -1,0 +1,8 @@
+﻿using _3ASystem.Domain.Entities.Operations;
+
+namespace _3ASystem.Application.Abstractions.Data.Repositories
+{
+	public interface IOperationRepository : IRepository<Operation, OperationId>
+    {
+    }
+}
