+++ src/3ASystem.Application/UseCases/Applications/Queries/GetApplications/GetApplicationsHandler.cs
@@ -1,6 +1,6 @@
-﻿using _3ASystem.Application.Abstractions.Messaging;
+﻿using _3ASystem.Application.Abstractions.Data.Repositories;
+using _3ASystem.Application.Abstractions.Messaging;
 using _3ASystem.Application.UseCases.Applications.Responses;
-using _3ASystem.Domain.Data.Repositories;
 using _3ASystem.Domain.Shared;
 
 namespace _3ASystem.Application.UseCases.Applications.Queries.GetApplications;
