+++ src/3ASystem.Application/UseCases/Modules/Commands/DeleteModule/DeleteModuleCommandHandle.cs
@@ -1,6 +1,6 @@
 ﻿using _3ASystem.Application.Abstractions.Data;
+using _3ASystem.Application.Abstractions.Data.Repositories;
 using _3ASystem.Application.Abstractions.Messaging;
-using _3ASystem.Domain.Data.Repositories;
 using _3ASystem.Domain.Entities.Applications;
 using _3ASystem.Domain.Entities.Modules;
 using _3ASystem.Domain.Shared;
