+++ src/3ASystem.Application/UseCases/Modules/Commands/UpdateModule/UpdateModuleCommandHandler.cs
@@ -1,9 +1,9 @@
 ﻿using _3ASystem.Application.Abstractions.Data;
+using _3ASystem.Application.Abstractions.Data.Repositories;
 using _3ASystem.Application.Abstractions.Messaging;
 using _3ASystem.Application.UseCases.Applications.Commands.UpdateApplication;
 using _3ASystem.Application.UseCases.Applications.Responses;
 using _3ASystem.Application.UseCases.Modules.Responses;
-using _3ASystem.Domain.Data.Repositories;
 using _3ASystem.Domain.Entities.Modules;
 using _3ASystem.Domain.Shared;
