+++ src/3ASystem.Application/UseCases/Modules/Queries/GetModules/GetModulesHandler.cs
@@ -1,7 +1,7 @@
-﻿using _3ASystem.Application.Abstractions.Messaging;
+﻿using _3ASystem.Application.Abstractions.Data.Repositories;
+using _3ASystem.Application.Abstractions.Messaging;
 using _3ASystem.Application.UseCases.Applications.Responses;
 using _3ASystem.Application.UseCases.Modules.Responses;
-using _3ASystem.Domain.Data.Repositories;
 using _3ASystem.Domain.Shared;
 
 namespace _3ASystem.Application.UseCases.Modules.Queries.GetModules;
