--- src/3ASystem.Domain/Data/Repositories/IFunctionalityRepository.cs
@@ -1,8 +1,0 @@
-﻿using _3ASystem.Domain.Entities.Functionalities;
-
-namespace _3ASystem.Domain.Data.Repositories
-{
-	public interface IFunctionalityRepository : IRepository<Functionality, FunctionalityId>
-	{
-	}
-}
