+++ src/3ASystem.Infrastructure/Data/Repositories/AppRepository.cs
@@ -1,4 +1,4 @@
-﻿using _3ASystem.Domain.Data.Repositories;
+﻿using _3ASystem.Application.Abstractions.Data.Repositories;
 using _3ASystem.Domain.Entities.Applications;
 using Microsoft.EntityFrameworkCore;
