+++ tests/3ASystem.Tests.Application/Application/Commands/DeleteApplicationCommandHandlerTests.cs
@@ -1,6 +1,6 @@
 ﻿using _3ASystem.Application.Abstractions.Data;
+using _3ASystem.Application.Abstractions.Data.Repositories;
 using _3ASystem.Application.UseCases.Applications.Commands.DeleteApplication;
-using _3ASystem.Domain.Data.Repositories;
 using _3ASystem.Domain.Entities.Applications;
 using _3ASystem.Domain.Shared;
 using FluentAssertions;
